Remove Mac storyboard
