+++ RepoZ.UI.Mac.Story/PopupViewController.cs
@@ -1,0 +1,48 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Foundation;
+using AppKit;
+
+namespace RepoZ.UI.Mac.Story
+{
+    public partial class PopupViewController : AppKit.NSViewController
+    {
+        #region Constructors
+
+        // Called when created from unmanaged code
+        public PopupViewController(IntPtr handle) : base(handle)
+        {
+            Initialize();
+        }
+
+        // Called when created directly from a XIB file
+        [Export("initWithCoder:")]
+        public PopupViewController(NSCoder coder) : base(coder)
+        {
+            Initialize();
+        }
+
+        // Call to load from the XIB/NIB file
+        public PopupViewController() : base("PopupView", NSBundle.MainBundle)
+        {
+            Initialize();
+        }
+
+        // Shared initialization code
+        void Initialize()
+        {
+        }
+
+        #endregion
+
+        //strongly typed view accessor
+        public new PopupView View
+        {
+            get
+            {
+                return (PopupView)base.View;
+            }
+        }
+    }
+}
